+++ backend/common/settings/structs.go
@@ -77,6 +77,7 @@ type Source struct {
 }
 
 type SourceConfig struct {
+	DenyByDefault    bool               `json:"denyByDefault"`           // deny access unless an "allow" access rule was specifically created.
 	Private          bool               `json:"private"`                 // designate as source as private -- currently just means no sharing permitted.
 	Disabled         bool               `json:"disabled"`                // disable the source, this is useful so you don't need to remove it from the config file
 	IndexingInterval uint32             `json:"indexingIntervalMinutes"` // optional manual overide interval in minutes to re-index the source