+++ frontend/src/views/Layout.vue
@@ -42,7 +42,7 @@ import { filesApi } from "@/api";
 import { state, getters, mutations } from "@/store";
 import { events } from "@/notify";
 import { generateRandomCode } from "@/utils/auth";
-import { shareOverrides } from "@/utils/constants";
+import { shareInfo } from "@/utils/constants";
 
 export default {
   name: "layout",
@@ -65,8 +65,8 @@ export default {
   },
   mounted() {
     window.addEventListener("resize", this.updateIsMobile);
-    if (shareOverrides.themeColor != "") {
-      document.documentElement.style.setProperty("--primaryColor", shareOverrides.themeColor);
+    if (shareInfo.themeColor != "") {
+      document.documentElement.style.setProperty("--primaryColor", shareInfo.themeColor);
     } else if (state.user.themeColor) {
       document.documentElement.style.setProperty("--primaryColor", state.user.themeColor);
     }